+++ compiler/__attribute__/hotpatch.h
@@ -1,0 +1,1 @@
+patchable_function_entry.h